Restructure project to move model inside model directory. Separate loss funciton logic to its own file. Create training and testing scripts.
